requirements.txt
ftplib==un
requests==2.31.0
easyocr==1.7.0
python-docx==0.8.11
PyPDF2==3.0.1
